+++ projects/demo-page/src/app/services/stackblitz.service.ts
@@ -1,9 +1,9 @@
-import { Inject, Injectable } from '@angular/core';
+import { APP_BASE_HREF } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
+import { Inject, Injectable } from '@angular/core';
 import stackBlitzSDK from '@stackblitz/sdk';
 import { Example } from '../examples';
 import { Utils } from '../utils';
-import { APP_BASE_HREF } from '@angular/common';
 
 function trimEndSlash(url: string): string {
   if (url[url.length - 1] === '/') {
@@ -14,14 +14,8 @@ function trimEndSlash(url: string): string {
 
 const templatePath = '/assets/stackblitz/';
 const templateFiles = [
-  '.editorconfig',
-  '.gitignore',
   'angular.json',
-  '.browserslistrc',
   'package.json',
-  'tsconfig.json',
-  'tsconfig.app.json',
-  'tslint.json',
   'src/index.html',
   'src/styles.scss',
   'src/polyfills.ts',
@@ -31,6 +25,10 @@ const replaceFilesPath = [
   'src/main.ts',
   'src/index.html',
 ];
+
+function getFilePath(example: Example, ext: 'ts' | 'css' | 'html') {
+  return `src/app/${example.name}.component.${ext}`;
+}
 
 
 @Injectable({
@@ -47,18 +45,22 @@ export class StackblitzService {
   }
 
   open(example: Example): void {
-    stackBlitzSDK.openProject({
-      files: this.getFiles(example),
-      title: 'ng-table-virtual-scroll | ' + example.title,
-      description: example.title,
-      template: 'angular-cli',
-      tags: ['angular', 'material', 'virtual scroll', 'table', 'ng-table-virtual-scroll'],
-      dependencies: {
-        '@angular/cdk': '^11.2.5',
-        '@angular/material': '^11.2.5',
-        'ng-table-virtual-scroll': '*'
+    stackBlitzSDK.openProject(
+      {
+        files: this.getFiles(example),
+        title: 'ng-table-virtual-scroll | ' + example.title,
+        description: example.title,
+        template: 'angular-cli',
+        dependencies: {
+          '@angular/cdk': '*',
+          '@angular/material': '*',
+          'ng-table-virtual-scroll': '*'
+        }
       },
-    });
+      {
+        openFile: [getFilePath(example, 'ts'), getFilePath(example, 'html')]
+      }
+    );
   }
 
   private setFiles(): void {
@@ -72,8 +74,8 @@ export class StackblitzService {
   }
 
   private getFiles(example: Example): { [path: string]: string } {
-    const exampleFiles = ['ts', 'css', 'html'].reduce((files, ext) => {
-      files[`src/app/${example.name}.component.${ext}`] = example[ext];
+    const exampleFiles = (['ts', 'css', 'html'] as const).reduce((files, ext) => {
+      files[getFilePath(example, ext)] = example[ext];
       return files;
     }, {});
     const replacedFiles = replaceFilesPath.reduce((files, path) => {